[Refactor] move `Result` into new file: `result.ts`
