+++ src/base/common/json.ts
@@ -1,4 +1,5 @@
-import { Result, tryOrDefault } from "src/base/common/error";
+import { tryOrDefault } from "src/base/common/error";
+import { Result } from "src/base/common/result";
 import { Arrays } from "src/base/common/utilities/array";
 import { Dictionary, Mutable, NonUndefined, Pair, isNumber, isObject, isString } from "src/base/common/utilities/type";
