+++ src/platform/configuration/common/configuration.ts
@@ -6,7 +6,7 @@ import { IConfigurationChangeEvent } from "src/platform/configuration/common/abs
 import { IService, createService } from "src/platform/instantiation/common/decorator";
 import { IRawConfigurationChangeEvent } from "src/platform/configuration/common/configurationRegistrant";
 import { URI } from "src/base/common/files/uri";
-import { AsyncResult, Result } from "src/base/common/error";
+import { AsyncResult, Result } from "src/base/common/result";
 
 export const APP_DIR_NAME = '.wisp';
 export const APP_CONFIG_NAME = 'app.config.json';
