+++ src/platform/configuration/electron/mainConfigurationService.ts
@@ -3,7 +3,7 @@ import { AbstractConfigurationService } from "src/platform/configuration/common/
 import { ILogService } from "src/base/common/logger";
 import { IInstantiationService } from "src/platform/instantiation/common/instantiation";
 import { IRegistrantService } from "src/platform/registrant/common/registrantService";
-import { AsyncResult } from "src/base/common/error";
+import { AsyncResult } from "src/base/common/result";
 
 export class MainConfigurationService extends AbstractConfigurationService {
