+++ src/platform/files/node/diskFileSystemProvider.ts
@@ -12,7 +12,8 @@ import { Watcher } from "src/platform/files/node/watcher";
 import { ILogService } from "src/base/common/logger";
 import { Emitter } from "src/base/common/event";
 import { IRawResourceChangeEvents, IWatcher } from "src/platform/files/common/watcher";
-import { errorToMessage, panic } from "src/base/common/error";
+import { errorToMessage } from "src/base/common/error";
+import { panic } from "src/base/common/result";
 import { Time, TimeUnit } from "src/base/common/date";
 
 export class DiskFileSystemProvider extends Disposable implements
