+++ src/workbench/services/fileTree/fileTreeCustomSorter.ts
@@ -1,6 +1,6 @@
 import { Time, TimeUnit } from "src/base/common/date";
 import { Disposable, IDisposable } from "src/base/common/dispose";
-import { AsyncResult, ok } from "src/base/common/error";
+import { AsyncResult, ok } from "src/base/common/result";
 import { DataBuffer } from "src/base/common/files/buffer";
 import { FileOperationError } from "src/base/common/files/file";
 import { URI } from "src/base/common/files/uri";
